requirements.txt
# Core dependencies
flask>=2.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
flask_cors
# PyTorch (pinned to avoid the weights_only issue)
torch>=2.0.0,<2.6.0
transformers>=4.20.0

# NLP engines
spacy


# Configuration
pyyaml>=6.0

# Optional: specific versions that work well together
# torch==2.5.1
# transformers==4.36.0
# stanza==1.6.0